requirements.txt
# --- Core Framework ---
Flask==2.2.5
Werkzeug==3.1.3
gunicorn==20.0.4
python-dotenv==1.1.1
flask-cors==5.0.0
flask-login==0.6.3
Flask-SQLAlchemy==3.1.1

# --- Auth / Security ---
bcrypt==4.1.2
email-validator==2.1.0

# --- Scheduler / Background Jobs ---
APScheduler==3.11.0

# --- Monitoring & Logging ---
sentry-sdk==2.19.2
blinker==1.9.0
elastic-apm==6.23.0

# --- Google OAuth & API ---
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0

# --- AI & Audio ---
openai==0.28.0
SpeechRecognition==3.14.1
pyttsx3==2.98

# --- PDF, Excel, and Document Tools ---
fpdf==1.7.2
openpyxl==3.1.5
XlsxWriter==3.2.0
xlrd==1.2.0

# --- GUI Support (Optional / Desktop Utilities) ---
PyQt5==5.15.11
PyQt5_sip==12.17.0

# --- Data Science / Plotting ---
 
matplotlib==3.10.3

# --- Database / ORM ---
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10

# --- Faker / Testing / Utilities ---
Faker==33.1.0
locust==2.37.12
pytest==5.4.2
pytest-cov==2.8.1
coverage==7.6.10
black


# --- Miscellaneous / Dependency Utilities ---
requests==2.32.4
appdirs==1.4.4
atomicwrites==1.4.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
click==8.1.8
colorama==0.4.6
gevent==24.11.1
greenlet==3.1.1
idna==2.10
Jinja2==3.0.3
MarkupSafe==3.0.2
more-itertools==10.5.0
packaging==24.2
pathspec==0.12.1
pluggy==0.13.1
py==1.11.0
pycparser==2.22
regex==2024.11.6
soupsieve==2.6
urllib3
whitenoise==6.8.2
uuid
wcwidth==0.2.13
wrapt==1.17.0
zope.event==5.0
zope.interface==7.2
ujson==5.9.0
cryptography==45.0.5
beautifulsoup4==4.9.1
numpy
pandas==2.3.1
Flask-Mail